chore: move article card to separate file
